fix no.15

+++ asafov.andrew/P3/matrixOutput.cpp
@@ -3,7 +3,7 @@
 void asafov::outputMatrix(std::ostream& out, const int* mtx, const unsigned long long n, const unsigned long long m)
 {
   out << n << ' ' << m << ' ';
-  for (int i = 0; i < n * m; i++) {
+  for (unsigned long long i = 0; i < n * m; i++) {
     out << mtx[i] << ' ';
   }
 }
